arca-prototype/requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.6
uvicorn[standard]>=0.32.0
playwright>=1.49.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
cryptography>=42.0.0
openai>=1.0.0
